Finished tidying everything up
Everything that worked before should be sorted and working fully now,
tested on local machine and it all runs fine. Tidied up a lot of the
JavaScript, Python and HTML. I've also downloaded a copy of the CSS and
JS for bootstrap and jQuery (since bootstrap requires it) and included
it into every HTML file. It shouldn't affect loading if unnecessary due
to caching so that should be fine.